+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -9,6 +9,9 @@ Notes](../../RELEASENOTES.md).
 * Fix `baggage` header not respecting the maximum length in some cases.
   ([#7061](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7061))
 
+* Improve efficiency of parsing of baggage and B3 propagation headers.
+  ([#7061](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7061))
+
 ## 1.15.2
 
 Released 2026-Apr-08